Move block::Header to primitives
